third_party/WebKit/LayoutTests/fast/dom/document-all-tags-expected.txt

Test that HTMLAllCollection.tags() matches IE behavior with too few arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.all.tags() threw exception TypeError: Failed to execute 'tags' on 'HTMLAllCollection': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
